+++ src/Email/Core/EmailExtension.php
@@ -10,11 +10,11 @@ declare(strict_types=1);
 namespace FreshAdvance\Invoice\Email\Core;
 
 use FreshAdvance\Invoice\Email\Settings\EmailSettingsInterface;
+use FreshAdvance\Invoice\Email\Traits\ServiceContainer;
 use FreshAdvance\Invoice\InvoiceData\DataType\InvoiceDataInterface;
 use FreshAdvance\Invoice\Pdf\InvoiceGeneratorInterface;
 use FreshAdvance\Invoice\Pdf\Service\FilenameCalculatorInterface;
 use FreshAdvance\Invoice\Service\Invoice;
-use FreshAdvance\Invoice\Traits\ServiceContainer;
 use OxidEsales\Eshop\Application\Model\Order;
 use OxidEsales\Eshop\Core\Email as ShopEmail;